Upgrade TeleportSystem & Implement Door Animation
Dynamic World Sample, Reusable Stage.prefab - Implement Door Animation and modify to cancel teleportation when the player exits the Stair layer.
DoorAnimationController - Deleted due to no use.
DoorController - Add enum of animator state and implement the ability to change animator parameters.
EndStageDoorController - A child class of DoorController, customized to fit the characteristics of the this gameObject.
FinalStageDoorController - A child class of DoorController, customized to fit the characteristics of the this gameObject.
StartStageDoorController - A child class of DoorController, customized to fit the characteristics of the this gameObject.
TeleportManager - Improved to cancel teleportation if the door opens during the teleportation attempt.

Files in Animation = Initiate Door Animation and Animation Controller.